devo7e file size reduction by converting u8 to unsigned (part 1)

+++ src/common.h
@@ -192,7 +192,7 @@ typedef enum {
 void PAGE_ShowInvalidStandardMixerDialog(void *guiObj);
 void STDMIXER_Preset();
 void STDMIXER_SetChannelOrderByProtocol();
-u8 STDMIXER_ValidateTraditionModel();
+unsigned STDMIXER_ValidateTraditionModel();
 const char *STDMIXER_ModeName(int mode);
 void STDMIXER_InitSwitches();
 void STDMIXER_SaveSwitches();
